avoid a build issue with non-threaded Win32 builds
The -DPERL_STATIC_SYMS was added in 9c293c15 as a replacement for a
kludge in 8803afc2 to fix SDBM built with BCC on Win32.

The failure in BCC was introduced in 17f28c40 which added the EXTCONST
to better support some VMS specific constructs (which appear to have
been removed since.)

A later patch may clean that up, but since the next dev release is
coming soon, I'll start with a minimal fix.